backend/requirements.txt

fastapi
uvicorn
openai
pyyaml
chromadb
python-dotenv
